+++ input/Areas.txt
@@ -1,0 +1,4 @@
+1,San Francisco
+2,New York City
+3,Detroit
+4,Dallas
