+++ src/storefront/services/SystemService.java
@@ -1,6 +1,13 @@
 package storefront.services;
 
+import java.util.ArrayList;
+
+import storefront.dao.AreaDAO;
+import storefront.dao.CustomerDAO;
+import storefront.dao.MachineDAO;
+import storefront.dao.ProductDAO;
 import storefront.dao.SystemDAO;
+import storefront.entities.Area;
 import storefront.entities.Customer;
 import storefront.entities.Machine;
 import storefront.entities.Product;
@@ -18,7 +25,12 @@ public class SystemService {
 		return instance;
 	}
 
-	SystemDAO dao = new SystemDAO();
+	private SystemDAO dao = new SystemDAO();
+	private ProductDAO pdao = new ProductDAO();
+	private MachineDAO mdao = new MachineDAO();
+	private CustomerDAO cdao = new CustomerDAO();
+	private AreaDAO adao = new AreaDAO();
+
 	
 	
 	public int commitPurchase(Customer c, Product p, Machine l, int dateStamp){
@@ -26,5 +38,26 @@ public class SystemService {
 	}
 	public int commitRequest(Customer c, Product p, Machine l, int dateStamp){
 		return dao.insertNewRequest(dateStamp, p.getProductID(), l.getMachineID(), c.getCustomerID());
+	}
+	
+	
+	
+	public ArrayList<Customer> retrieveAllCustomers() {
+		return cdao.readAllCustomersFromDB();
+	}
+	
+	public ArrayList<Machine> retrieveAllMachines() {
+		return mdao.readAllMachinesFromDB();
+	}
+	
+	public ArrayList<Product> retrieveAllProducts() {
+		ArrayList<Product> productList = new ArrayList<Product>();
+		//TODO
+		productList = pdao.readAllProductsFromDB();
+		return productList;
+	}
+	
+	public ArrayList<Area> retrieveAllAreas() {
+		return adao.readAllAreasFromDB();
 	}
 }
